feat: products structure

+++ server/src/users/users.controller.ts
@@ -7,6 +7,7 @@ import {
   Param,
   Delete,
   ConflictException,
+  NotFoundException,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
@@ -17,31 +18,43 @@ export class UsersController {
   constructor(private readonly usersService: UsersService) {}
 
   @Post()
-  create(@Body() createUserDto: CreateUserDto) {
-    const oldUser = this.usersService.findOneByEmail(createUserDto.email);
+  async create(@Body() createUserDto: CreateUserDto) {
+    const user = await this.usersService.findOneByEmail(createUserDto.email);
 
-    if (oldUser) throw new ConflictException(`This email is already in use`);
+    if (user) throw new ConflictException(`This email is already in use`);
 
     return this.usersService.create(createUserDto);
   }
 
   @Get()
-  findAll() {
+  async findAll() {
     return this.usersService.findAll();
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.usersService.findOne(id);
+  async findOne(@Param('id') id: string) {
+    const user = await this.usersService.findOne(id);
+
+    if (!user) throw new NotFoundException(`User with id: '${id}' not found`);
+
+    return user;
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+  async update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+    const user = await this.usersService.findOne(id);
+
+    if (!user) throw new NotFoundException(`User with id: '${id}' not found`);
+
     return this.usersService.update(id, updateUserDto);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  async remove(@Param('id') id: string) {
+    const user = await this.usersService.findOne(id);
+
+    if (!user) throw new NotFoundException(`User with id: '${id}' not found`);
+
     return this.usersService.remove(id);
   }
 }
